Use a more specific name
We might have a storage root, etc. Also, the graph and the semantics are not identical, as multiple workflows might have a semantic (but not operational) interaction, so then we'd care about a semantic root! Plus it makes find-usages easier on the IDE when we have a more specific name.