Disallow IO maps on `AbstractMacro`
I tried to break this down a reasonable amount, but I also want the tests passing at each commit. In principle, I could duplicate code instead of simply moving it to break this apart a bit further, but in practice I'm not expecting anyone to actually review this PR so that seems like needless effort. Here, I

* Move `inputs/outputs_map` from `Composite` down to `Workflow`
* Disable automatic IO generation on `AbstractMacro` -- you _need_ to define it by labels and the graph creator signature/returns
* Add IO previewing and class-level stuff analogous to `AbstractFunction` onto `AbstractMacro` now that its IO can't be remapped
* Update tests and examples accordingly
* Finally, re-write the while- and for-loop meta nodes, which depended heavily on the presence of maps

Note that although the for- and while-loops are totally different under the hood, the only user-facing changes are (a) you can only wrap nodes that are importable, and (b) the IO maps for the while-loop are mandatory instead of optional.