:fire: :bug: make sure for-loops re-set their value links like macros (#373)
Honestly, I basically just forgot that for-loops are not actually macros. Ideally, this behaviour would come from a shared parent class, like `StaticComposite` analogous to `StaticNode`. Unfortunately, that gives circular imports because the creator object is on `Composite`. Once `Workflow` and `Macro` do not share access to the creator, such an abstraction will be possible. In the meantime, I just tolerate the once-duplicated code.