[patch] Fail to cache when equality comparison throws an exception
E.g. when comparing a numpy array to None. We continue greedily caching, but if it just doesn't work for some reason, the fallback is to simply not cache. Since the current and cached inputs (along with whether or not they generate a cache hit) is always available to the user, I think we still leave enough data for people to debug cache misses on their own.