[patch] Remove the need to double-decorate
If you're making something a dataclass node, you can now pass in a dataclass-compliant class that is not yet a dataclass and we'll take care of turning it into one for you. That means you only need `@Workflow.wrap.as_dataclass_node` and not `@Workflow.wrap.as_dataclass_node;@dataclass`. Per Joerg's wishlist.