[minor] Shift responsibility (#430)
* Remove unused property

* Shorten variable name

* Refactor: rename method

* Call base method and remove specialist

* Condense has_contents

* Remove unused property

* Move methods from Node to StorageInterface

* Have StorageInterface.load just return the new instance

* Break dependence on Node.working_directory

* Refactor: rename obj to node in the storage module

Since it is the interface for Node instances

* :bug: pass argument

* Remove tidy method

* Clean up using pathlib

* Rely on pathlib

This is almost exclusively because `pyiron_snippets.files.DirectoryObject` has no option to avoid the `self.create()` call at instantiation. That means if I want to peek at stuff, I need to create then clean it, which is killer slow. I'm very open to re-introducing `files` dependence again in the future, I just don't have the time to implement the necessary fixes, with test, and go through the release cycle. This is all under-the-hood, so should be safely changeable with a `pyiron_workflow` patch change later anyhow.

* Remove the post-setup cleanup

No longer necessary now that storage isn't creating directories willy nilly

* Simplify and remove unused

* Allow semantic objects to express as pathlib.Path

* Leverage new semantic path for storage

* Remove unneeded exception

* Remove unused imports

* Use a generator in the storage module for available backends

* Test loading failure

* Fix docstring

* Format black

* Remove unused import

* Don't overload builtin name

---------

Co-authored-by: pyiron-runner <pyiron@mpie.de>