[minor] Replace `HasToDict` mixin with `HasStateDisplay` mixin (#435)
* [minor] Replace `HasToDict` mixin with `HasStateDisplay` mixin

Functionality is similar -- the point is to get a nice JSON representation for child classes. But the public interface is different (hence minor flag), and in particular we don't use `to_dict` which might interfere with other pyiron activities. There's also some changes to exactly the display looks like, but since it's just a nice/useful thing for humans to look at, this is non-critical and can be improved with patches as needed.

* Remove unused import