[patch] Bump development status to Beta
The filesystem interaction is stable and robust (i.e. recovery files get written when things go wrong), which has the knock-on effect that you can scale workflows to remote processes. Support for individual nodes running on remote processes beyond the lifetime of the parent python process is pretty rough, but also there.