`mypy` semantics (#538)
* Initialize _label to a string

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Hint the delimiter

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Make SemanticParent a Generic

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Purge `ParentMost`

If subclasses of `Semantic` want to limit their `parent` attribute beyond the standard requirement that it be a `SemanticParent`, they can handle that by overriding the `parent` setter and getter. The only place this was used was in `Workflow`, and so such handling is now exactly the case.

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Update comment

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Use generic type

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Don't use generic in static method

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Jump through mypy hoops

It doesn't recognize the __set__ for fset methods on the property, so my usual routes for super'ing the setter are failing. This is annoying, but I don't see it being particularly harmful as the method is private.

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Remove unused import

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Add dev note

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

---------

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>