`mypy` finishing touches (#562)
* Make class "property" a plain method

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Refactor to non-None classvar

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Explicitly cast to tuple

For the sake of the name generator

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Silence mypy

It is upset about the hinting `list[hint]`, because `hint` is variable. We would be able to verify that it is a type at static analysis time, but since it comes from the body node class -- which is unknown until runtime -- it is impossible to say _which_ type.

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Uniformly give and ignore classfactory hints

At a minimum, getting mypy to parse these correctly requires more rigorous hinting in pyiron_snippets.factory. But actually, since the classfactory allows the parent class to be specified with _multiple bases_, I'm not even 100% sure we'd ever be able to get a single type variable that could do the trick universally. In any case, for now kick the can don't the road and always hint what you know is true, then tell mypy to not worry about it.

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Add some hints to preview

Albeit pretty relaxed ones

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Add a return hint to Runnable.__init__

To get mypy to parse the body of the function

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Break loop into a method

Mypy didn't like parsing the zip variable when it could be inputs or outputs (even though both inherit from the relevant DataIO in this case), but using a separate method is functionally equivalent and mypy can get a better grasp of the type values.

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

---------

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>