Reparent AmbiguousOutputError onto AttributeError (#595)
* Reparent AmbiguousOutputError onto AttributeError

Just conceptually errors being raised from `__getattr__` ought to be this way anyhow, not `ValueError`. But more importantly, raising a `ValueError` instead of an `AttributeError` meant that the builtin `hasattr` method couldn't properly return `False` when the lookup failed due to not having an unambiguous `.channel` attribute to which to delegate the lookup.

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Lint

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

---------

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>