Emit signals when running a child directly
Needs a test to ensure this "push" behaviour stays working, and I want to explore recovery files a little bit. Still, even if the recovery file is less graceful under this condition, I feel that would be a fair price to pay for going into the innards to run something.

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>