Remove unreached block
`self` is never in `self._parent.children` because we need that to either be `self.label` or `self._parent.children.values()`. However, making this fix breaks the tests for parallel processing, and we anyhow guarantee that the child is parentless a few lines later inside `self._parent.add_child(self)`. TBH, I don't totally understand why fixing this clause so it actually runs breaks things, but since we're guaranteed to get an exception if we need to remove the parent and haven't, I'm just going to remove it.

Thanks to mypy for catching that the item type being search for and container's item type didn't match.

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>