[patch] Satisfy updated mypy and configure codecov (#653)
* Specify valued types for None-initialized attribute

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Remove unreached block

`self` is never in `self._parent.children` because we need that to either be `self.label` or `self._parent.children.values()`. However, making this fix breaks the tests for parallel processing, and we anyhow guarantee that the child is parentless a few lines later inside `self._parent.add_child(self)`. TBH, I don't totally understand why fixing this clause so it actually runs breaks things, but since we're guaranteed to get an exception if we need to remove the parent and haven't, I'm just going to remove it.

Thanks to mypy for catching that the item type being search for and container's item type didn't match.

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* cast away recursive type var problem

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Replace HasGenericChannel

With something simpler and more concrete

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Use the more concrete type

In all the injected outputs. Now in macro.py:270, the `node.channel` should indeed by guaranteed to be an `OutputDataWithInjection` and thus have a `value_receiver`, instead of merely being guaranteed to be a `Channel`

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Adjust coverage behaviour

Including codecov configuration

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

---------

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>