Restrict input data to have at most one connection (#673)
* Disallow multiple data input connections

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Create syntactic sugar for input panels

So that we can automatically disconnect existing data input connections when assigning a new connections

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Update deepdive

Removing the "input is prioritized" section and replacing it with a demo of how each data input can only have one connections

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Amend the node docstring

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Remove the manual while-loop test

Allowing a single input connection expressly destroys this concept. We already replaced the functionality with a `While` node.

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

---------

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>