Wrap executorlib executors
To exploit the new caching interface so that nodes can rely on their running state and lexical path to access previously executed results.

Locally with the SingleNodeExecutor everything is looking good, but that doesn't natively support terminating the process that submit the job. I'd like to play around with this using the SlurmClusterExecutor on the cluster before making further changes.

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>