Break apart data tree logic
To remove the local import on Workflow and have parented trees run their own, simpler logic when possible. There is quite a bit of duplicate code here, but let's commit something working and then worry about making it prettier. I used Pycharm's Junie for the first crack at this. I'm not sure it actually saved time, but at least it came close to writing something that kept the tests passing...

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>