Remove Lexical.parent_type
Lexical is still generic on the type hint, but only for type hinting not on the actual parent class object. We accomplish this by shifting the entire responsibility of typing/cylcicity/parentage over to the parent -- so LexicalParent still needs to explicitly know its child type, but this way it's a unidirectional dependency and we can get rid of the local import of `Composite` inside `Node`.

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>