Better rerun (#723)
* Isolate subtests

And fix bug in the second test; we accidentally suppressed it when the subtests were dependent.

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Move rerun changes to _before_run

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Consolidate callbacks

So we never run into the case where the future no longer exists

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Return encountered futures

And fail if there is a live future but the user hard-requested a re-run. In this case the node might not just be in a "running" state but might literally still be computing something in another thread/process!

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Decreaser line lengths

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Make a more specific error

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

---------

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>