Trim docstrings (#730)
* Remove promises from docstrings

These are declarations about behaviour that are decoupled from executions of the behaviour they describe. As such, I don't trust them to stay true. Just cut them all. This information is carried in the docstring examples, the deep dive notebook, or the test suite -- or it isn't real information.

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* De-duplicate docstrings

By migrating subsections in the class-level docstring to the right spot, or removing them if there's already something there. PyCharm parses the `Attributes` and highlights things there that are not attributes (although sadly the reverse is not true), so this shuffle makes those highlights more trustworthy.

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

---------

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>